Add NXP i.MX RT1060 BootROM
This is one of NXP's "Crossover" SoCs, more powerful than a typical
microcontroller but not powerful enough to run Linux. It has a Cortex-M7
clocked at 600MHz and 1MiB of onboard SRAM.

This BootROM was extracted from a MIMXRT1061DVL6A part inside a Google
Stadia controller. The file's contents are mapped at 0x0020_0000.